% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survfit.R
\name{export_survfit}
\alias{export_survfit}
\alias{survfit}
\title{\code{survfit} function imported from \code{survival}}
\description{
This function is imported from the \code{survival} package. See
\code{\link[survival]{survfit}}.
}
