% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{extand}
\alias{extand}
\title{Extend a \code{qris} to a specified range of $\tau$ or $t_0$ values.}
\usage{
extand(x, t0s = NULL, Qs = NULL, nB = NULL, vari = NULL)
}
\arguments{
\item{x}{is an qris object or a data.frame returned by plot.qris}

\item{t0s}{is a vector of range of t0 to plot; when not specified, the default value is from 0 to presently defined \eqn{t_0}}

\item{Qs}{is a vector of range of Q to plot; when not specified, the default value is from 5\% to presently defined \eqn{Q}}

\item{nB}{is the number of multiplier bootstrapping for standard error estimation.}

\item{vari}{is a character string to choose variables to draw the regression coefficient.}
}
\description{
Extend a \code{qris} to a specified range of $\tau$ or $t_0$ values.
}
