% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ismb_est.R
\name{ismb_est}
\alias{ismb_est}
\title{Quantile regression estimator using induced smoothing approach}
\usage{
ismb_est(Z, nc, covariate, D, t_0, Q, ne)
}
\arguments{
\item{Z}{A vector of subjects' event time}

\item{nc}{Number of covariate to estimate}

\item{covariate}{A matrix of covariate (dimension number of subjects(n) x number of covariate(nc))}

\item{D}{A vector of subjects' censored indicator (1 = not_censored data, 0 = censored data)}

\item{t_0}{Followup time or base time adjustment (minimum t_0 is 0, and it means no base time adjustment)}

\item{Q}{Quantile that user want to estimate (scale : 0 ~ 1)}

\item{ne}{Number of eta generation when estimate standard error of estimator}
}
\value{
A nc by 2 matrix (First column is beta and second column is standard error of betas)
}
\description{
ismb_est function will calculate qunatile regression parameters "beta" and standard error of "beta" using induced smoothing approach.
}
\examples{
    covar = as.matrix(data[,1:12])
    ismb_est(data$survTime, 12, covar, data$event, 2, 0.5 ,100)
}
