% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qris_pkg.R
\docType{package}
\name{qris-package}
\alias{qris-package}
\alias{qris-packages}
\title{qris: Quantile Regression Model for Residual Lifetime Using an Induced Smoothing Approach}
\description{
This package offers a collection of functions to fit quantiles regression models for censored residual lifetimes. It provides various options for regression parameters estimation: the induced smoothing approach (smooth), and L1-minimization (non-smooth).  It also implements the estimation methods for standard errors of the regression parameters estimates based on an efficient partial multiplier bootstrap method and robust sandwich estimator. Furthermore, a simultaneous procedure of estimating regression parameters and their standard errors via an iterative updating procedure is implemented (iterative). See Kim, K. (2022) "Smoothed quantile regression for censored residual life", <arXiv:2205.00413>.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Kyuhyun07/qris}
}

}
\author{
\strong{Maintainer}: Kyu Hyun Kim \email{kyuhyunkim07@yonsei.ac.kr}

Authors:
\itemize{
  \item Sangwook Kang
  \item Sy Han Chiou
}

}
