% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qris.R
\name{qris.control}
\alias{qris.control}
\title{Auxiliary for Controlling qris}
\usage{
qris.control(maxiter = 10, tol = 0.001, trace = FALSE)
}
\arguments{
\item{maxiter}{max number of iteration.}

\item{tol}{tolerance of convergence}

\item{trace}{a binary variable, determine whether to save output for each iteration.}
}
\value{
A list with the arguments as components.
}
\description{
Auxiliary function as user interface for \code{qris} fitting.
}
\details{
When \code{trace} is TRUE, output for each iteration is printed to the screen.
}
\seealso{
\code{\link{qris}}
}
