% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrismb.R
\name{qrismb.control}
\alias{qrismb.control}
\title{Auxiliary for Controlling qrismb}
\usage{
qrismb.control(maxiter = 10, tol = 0.001, trace = FALSE)
}
\arguments{
\item{maxiter}{max number of iteration.}

\item{tol}{tolerance of convergence}

\item{trace}{a binary variable, determine whether to display output for each iteration.}
}
\value{
A list with the arguments as components.
}
\description{
Auxiliary function as user interface for \code{qrismb} fitting.
}
\details{
When \code{trace} is TRUE, output for each iteration is printed to the screen.
}
\seealso{
\code{\link{qrismb}}
}
