% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{residuals.qris}
\alias{residuals.qris}
\title{Residuals for Quantile Regression Model Fitted on Residual life}
\usage{
\method{residuals}{qris}(object, newdata, ...)
}
\arguments{
\item{object}{is a qris object}

\item{newdata}{is a data frame for an optional new data to do predictions.
If omitted, the fitted values based on the original data and fit will be returned.}

\item{...}{for future extension}
}
\value{
A vector of residual
}
\description{
Residual based on fitted quantile regression model
}
