% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrismb.R
\name{ghat}
\alias{ghat}
\title{Estimate Kaplan Meier estimate of the survival function of the censoring time C
ghat}
\usage{
ghat(T, censor, wgt = 1)
}
\arguments{
\item{T}{is a vector of observed time, which is minimum of failure time and censored time}

\item{wgt}{is a vector of weight}

\item{event}{is a vector of censoring indicator (not censored = 1, censored = 0)}
}
\value{
An object of class "\code{ghat}" estimate KM estimator of survival function of
The \code{ghat} object is a list containing at least the following components:
\describe{
  \item{deathtime}{the observed time}
  \item{ndeath}{a vector of number of subject who experienced event at deathtime}
  \item{nrisk}{a vector of number of subject who are possible to experience event at deathtime}
  \item{survp}{a vector of survival probability at deathtime}
  }
}
\description{
Estimate Kaplan Meier estimate of the survival function of the censoring time C
ghat
}
