% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.qrismb}
\alias{plot.qrismb}
\title{Draw 95\% confidence interval by a quantile regression estimator of residual lifetime from survival data}
\usage{
\method{plot}{qrismb}(object, type = 2, t0s = c(0, t0), Qs = c(0.05, Q), ne = 100, ...)
}
\arguments{
\item{object}{is an qrismb object}

\item{type}{is a indicator that which parameter designated to x value (1 = t0, 2 = quantile); when not specified, the default value is 2.}

\item{t0s}{is a vector of range of t0 to plot; when not specified, the default value is from 0 to presently defined \eqn{t_0}}

\item{Qs}{is a vector of range of Q to plot; when not specified, the default value is from 5\% to presently defined \eqn{Q}}

\item{ne}{is the number of multiplier bootstrapping
when not specified, multiplier bootstrap will be carried out with the \code{ne} specified in \code{object};
when \eqn{ne = 0}, only the point estimates will be plotted;
when \eqn{ne > 1}, both the point estimates and the 95\% Wald CI will be plotted.}
}
\description{
Draw 95\% confidence interval by a quantile regression estimator of residual lifetime from survival data
}
