% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qris_pkg.R
\docType{package}
\name{qris-package}
\alias{qris-package}
\alias{_PACKAGE}
\alias{qris-packages}
\title{qris: Quantile regression using induced smoothing approach}
\description{
The package offers ...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Kyuhyun07/qris}
}

}
\author{
\strong{Maintainer}: Kyu Hyun Kim \email{kyuhyunkim07@yonsei.ac.kr}

Authors:
\itemize{
  \item Sangwook Kang
  \item Sy Han Chiou
}

}
