% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrismb_pkg.R
\docType{package}
\name{qrismb-package}
\alias{qrismb-package}
\alias{_PACKAGE}
\alias{qrismb-packages}
\title{qrismb: Quantile regression using induced smoothing approach}
\description{
The package offers ...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Kyuhyun07/qrismb}
}

}
\author{
\strong{Maintainer}: Kyu Hyun Kim \email{kyuhyunkim07@yonsei.ac.kr}

Authors:
\itemize{
  \item Sangwook Kang
  \item Sy Han Chiou
}

}
