% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qrismb.R
\name{qrismb}
\alias{qrismb}
\title{Estimate a quantile regression estimator of residual lifetime from survival data}
\usage{
qrismb(
  formula,
  data,
  t0 = 0,
  Q = 0.5,
  ne = 100,
  method = c("smooth", "iterative", "nonsmooth"),
  init = c("rq", "noeffect", "userinput"),
  userinit
)
}
\arguments{
\item{formula}{a formula expression, of the form \code{response ~ predictors}.
The \code{response} is a \code{Surv} object with right censoring.}

\item{data}{an optional data.frame in which to interpret the variables occurring in the \code{formula}.}

\item{t0}{is the followup time(or basetime of analysis)}

\item{Q}{is the quantile}

\item{ne}{is number of multiplier bootstrapping for V matrix estimation}

\item{method}{is an option for specifying the methods of parameters and and standard errors estimation
("smooth" is default in which parameters estimates and their standard errors ore obtained via induced smoothed estimating equations and partial multiplier bootstrapping, respectively.
"nonsmooth" uses a L1-minimization method for non-smooth object functions in coefficient estimation and a full multiplier bootstrappong in standard errors estimation.
"iterative" simultaneously estimates parameters and their standard errors based the iterative updates for parameter estimates and their standard errors based on induced smoothed estmating equtions and partial multiplier bootstrapping)}

\item{init}{is an option for specifying the initial values of the parameters estimates
("rq" is default in which the estimates from the non-smooth counterpart is specified,
"noeffect" specifies no covariate effect except for intercept c(1,0,0, ...)
"userinput" specifies a user defined vector as an initial value)}

\item{userinit}{is a user defined initial guess vector if user chooses "userinput" as init parameter.}
}
\value{
An object of class "\code{qrismb}" contains model fitting results.
The \code{qrismb} object is a list containing at least the following components:
\describe{
  \item{coefficient}{a vector of point estimates}
  \item{stderr}{a vector of standard error of point estiamtes}
  \item{iterno}{a number of itertation until convergence (only for iterative procedure)}
  }
}
\description{
Using three estimation methods
1. L1-minimization(non-smooth estimating equation)
2. Induced smoothing approach (smooth estimating equation)
3. Iterative procedure with induced smoothing approach (smooth estimating equation)
}
\examples{
## #########################################
## Simulated data
## #########################################
data.gen <- function(n) {
    r0 <- .2 * sqrt(log(2))
    r1 <- .1 * sqrt(log(2))
    dat <- data.frame(censoring = runif(n, 0, 24.35),
                      Time0 = sqrt(-log(1 - runif(n))),
                      X = rbinom(n, 1, .5))
    dat$Time0 <- ifelse(dat$X > 0, dat$Time0 / r1, dat$Time0 / r0)
    dat$Time <- pmin(dat$Time0, dat$censoring)
    dat$status <- 1 * (dat$Time0 < dat$censoring)
    subset(dat, select = c(Time, status, X))
}

dat <- data.gen(200)

fm <- Surv(Time, status) ~ X
fit1 <- qrismb(fm, data = dat, t0 = 1, Q = 0.5, ne = 200, "smooth", "rq")
fit2 <- qrismb(fm, data = dat, t0 = 1, Q = 0.5, ne = 200, "nonsmooth", "noeffect")
fit3 <- qrismb(fm, data = dat, t0 = 1, Q = 0.5, ne = 200, "iterative", "userinput", c(2,1))

summary(fit1)
summary(fit2)
summary(fit3)

## #########################################
## Real data application
## #########################################
data(cancer, package = "survival")
lung2 <- subset(lung, select = c(time, status, age, sex))
## tidy up the data
lung2$status <- lung2$status - 1
lung2$sex <- lung2$sex - 1

fm <- Surv(time, status) ~ age + sex
fit1 <- qrismb(fm, data = lung2, t0 = 0, Q = 0.5, ne = 200, "iterative")
fit2 <- qrismb(fm, data = lung2, t0 = 30, Q = 0.5, ne = 200, "nonsmooth", "userinput", c(1,0,1))
fit3 <- qrismb(fm, data = lung2, t0 = 100, Q = 0.5, ne = 200,"smooth", "noeffect")

summary(fit1)
summary(fit2)
summary(fit3)

}
