% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{residuals.qris}
\alias{residuals.qris}
\title{Residuals for Quantile Regression Model Fitted on Residual life}
\usage{
\method{residuals}{qris}(object, newdata, ...)
}
\value{
A vector of residual
}
\description{
Residual based on fitted quantile regression model
}
